+++ src/main/java/com/guba/spring/speakappbackend/database/models/Image.java
