+++ src/main/java/com/guba/spring/speakappbackend/database/models/Task.java
@@ -1,7 +1,5 @@
 package com.guba.spring.speakappbackend.models;
 
-import com.guba.spring.speakappbackend.converters.CategoryJpaConverter;
-import com.guba.spring.speakappbackend.enums.Category;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
