+++ src/main/java/com/guba/spring/speakappbackend/database/repositories/ImageTempRepository.java
