+++ src/main/java/com/guba/spring/speakappbackend/web/schemas/GenerateExerciseResponse.java
@@ -1,4 +1,4 @@
-package com.guba.spring.speakappbackend.schemas;
+package com.guba.spring.speakappbackend.web.schemas;
 
 import com.guba.spring.speakappbackend.models.Exercise;
 import lombok.AllArgsConstructor;
