+++ src/main/java/com/guba/spring/speakappbackend/web/schemas/ImageDTO.java
@@ -1,4 +1,4 @@
-package com.guba.spring.speakappbackend.schemas;
+package com.guba.spring.speakappbackend.web.schemas;
 
 import com.guba.spring.speakappbackend.models.Image;
 import lombok.AllArgsConstructor;
