+++ src/main/java/com/guba/spring/speakappbackend/web/schemas/MultipartFileDTO.java
@@ -1,4 +1,4 @@
-package com.guba.spring.speakappbackend.schemas;
+package com.guba.spring.speakappbackend.web.schemas;
 
 
 import lombok.AllArgsConstructor;
